regression and classification dist classes
- All distributions are now a subclass of the intermediate classes RegressionDistn or ClassificationDistn, which implement the appropriate predict() methods instead of having those in the API. Anything that is a RegressionDist needs to have mean() implemented, while anything that is a ClassificationDist needs to have class_probs() implemented.

Also moved manifold() out of distsn.py into its own top-level file
