+++ ngboost/__init__.py
@@ -1,4 +1,4 @@
 from .ngboost import NGBoost
 from .api import NGBClassifier, NGBRegressor, NGBSurvival
 
-__version__ = "0.1.5dev"
+__version__ = "0.2.0dev"
