+++ ngboost/distns/__init__.py
@@ -1,4 +1,4 @@
-from .distn import Distn, manifold
+from .distn import RegressionDistn, ClassificationDistn
 from .normal import Normal, NormalFixedVar
 from .multivariate_normal import MultivariateNormal
 from .lognormal import LogNormal
